requirements.txt
attrs==25.1.0
cattrs==24.1.2
coremltools==8.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
matplotlib-venn==1.1.2
mdit-py-plugins==0.4.2
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.2
pandas-datareader==0.10.0
pandas-gbq==0.28.0
pandas-stubs==2.2.2.240909
pillow==11.1.0
protobuf==4.25.6
psutil==5.9.5
py-cpuinfo==9.0.0
pyaml==25.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
torchsummary==1.5.1
tornado==6.4.2
tqdm==4.67.1
ultralytics==8.3.85
